Move the writer stage's utilities to its own module
